Type Casting Correctness
Major TODO, but this is just a bunch of places I fixed this round.

+++ src/jaudio/dummyrom.c
@@ -69,7 +69,7 @@ void ARAlloc2(u32)
  */
 void* ARAllocFull(u32* outSize)
 {
-	u32 freeSize = aram_hp.length - ((int)aram_hp.current - (int)aram_hp.base);
+	u32 freeSize = aram_hp.length - (aram_hp.current - aram_hp.base);
 
 	void* alloc = Nas_HeapAlloc(&aram_hp, freeSize - 32);
 	*outSize    = freeSize - 32;
